Upload test Typescript version
