+++ package.json
@@ -1,25 +1,23 @@
 {
-  "name": "portfolio",
+  "name": "portfolio-ts",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "boostrap": "^2.0.0",
-    "bootstrap": "^4.5.2",
-    "css.gg": "^2.0.0",
-    "node-sass": "^4.14.1",
+    "@types/jest": "^24.9.1",
+    "@types/node": "^12.12.54",
+    "@types/react": "^16.9.46",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-router-dom": "^5.1.5",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
-    "react-helmet": "^6.1.0",
-    "react-native": "^0.63.2",
-    "react-responsive-carousel": "^3.2.9",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
-    "react-spring": "^8.0.27",
-    "styled-components": "^5.1.1"
+    "reactjs-localstorage": "^1.0.1",
+    "typescript": "^3.7.5"
   },
   "scripts": {
     "start": "react-scripts start",
