+++ src/App.css
@@ -3,12 +3,14 @@
 }
 
 .App-logo {
-  height: 60vmin;
+  height: 40vmin;
   pointer-events: none;
 }
 
-.background-dark {
-  
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
 }
 
 .App-header {
