+++ src/Components/Content.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import {Container, Row, Col} from 'react-bootstrap';
+
+const Content: React.FC = (props) => {
+    
+    return (
+        <Container fluid={true}>
+            <Row className="justify-content-center">
+                <Col md={8}>
+                    {props.children}
+                </Col>
+            </Row>
+        </Container>
+    )
+}
+
+export default Content;
