--- src/Components/Footer.modules.css
@@ -1,6 +1,0 @@
-.footer {
-    position: absolute;
-    bottom: 0;
-    width: 100%;
-    height: 4rem;
-}
