--- src/Pages/ContactPage.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function ContactPage(props) {
-    
-    return(
-        <p>ContactPage Works!</p>
-    );
-}
-
-export default ContactPage;
