+++ src/Pages/ContactPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import {Page} from '../Components/AppProps';
+
+const ContactPage: React.FC<Page> = (props: Page) => {
+
+    return (
+        <p>ContactPage Works!</p>
+    );
+}
+
+export default ContactPage;
