--- src/Pages/ProjectsPage.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Hero from '../Components/Hero';
-import Carousel from '../Components/Carousel';
-
-
-function ProjectsPage(props) {
-    
-    return(
-        <div>
-            <Hero title={props.title} subTitle={props.subTitle} text={props.text} />
-            <Carousel />
-        </div>
-    );
-}
-
-export default ProjectsPage;
