+++ src/index.css
@@ -1,6 +1,5 @@
 body {
   margin: 0;
-  background-color: darkgrey;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
@@ -8,46 +7,7 @@ body {
   -moz-osx-font-smoothing: grayscale;
 }
 
-#root {
-  margin: 0;
-  background-color: darkgrey;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-.d-carousel {
-  margin: 10px;
-
-}
-
-.d-card {
-  margin: 20px;
-  min-height: 100vh;
-}
-
-.d-card-img {
-  border-radius: 10px;
-  /* width: 300px;
-  height: 480px; */
-  box-shadow: 0px 0px 3px 1px #ccc;
-}
-
-
-.d-card-info {
-  margin-top: 10px;
-  min-height: 100px;
-}
-
-
-.d-card-title {
-  font-size: 24px;
-  margin: 0px;
-}
-
-.d-card-sub-title {
-  font-size: 16px;
-  margin: 0px;
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
 }
