--- src/styles.css
@@ -1,9 +1,0 @@
-.light-mode {
-    background-color: #f0f0f0;
-    color: #333;
-}
-
-.dark-mode {
-    background-color: #1a1919;
-    color: #999;
-}
